Added support of chip sizes more than 4GB

+++ qt/parallel_chip_info.h
@@ -64,7 +64,7 @@ private:
     } StmParams;
 
     QByteArray halConf;
-    uint32_t params[CHIP_PARAM_NUM] = {};
+    quint64 params[CHIP_PARAM_NUM] = {};
 
     void chipInfoToStmParams(StmParams *stmParams);
 
@@ -72,8 +72,8 @@ public:
     ParallelChipInfo();
     virtual ~ParallelChipInfo();
     const QByteArray &getHalConf() override;
-    uint32_t getParam(uint32_t num) override;
-    int setParam(uint32_t num, uint32_t value) override;
+    quint64 getParam(uint32_t num) override;
+    int setParam(uint32_t num, quint64 value) override;
 };
 
 #endif // PARALLEL_CHIP_INFO_H
