+++ qt/writer.cpp
@@ -18,7 +18,7 @@ Writer::~Writer()
 }
 
 void Writer::init(SerialPort *serialPort, QVector<uint8_t> *buf,
-    uint32_t addr, uint32_t len, uint32_t pageSize, bool skipBB, bool incSpare,
+    quint64 addr, quint64 len, uint32_t pageSize, bool skipBB, bool incSpare,
     bool enableHwEcc, uint8_t startCmd, uint8_t dataCmd, uint8_t endCmd)
 {
     this->serialPort = serialPort;
